Decouple Storage and Node Actors - Removed Singleton Pattern - Added forwarding to Node
